Unmapped options should be added to the mapping with no processing.
In order to do this correctly custom argument parsing had to
be added to the library i.e. we no longer rely on minimist.
This is due to the fact that minimist is not able to remember
what the format the options were in when they were originally
passed to it.

Closes #16.